Few changes
slightly changed the code + added a check for the presence of resources
