+++ saturnine/__main__.py
@@ -1,6 +1,6 @@
 from sdk_server.app import run_http_server
 from game_server import GameServer;
-from game_server.handlers import auth, ping, scene, avatar, entity, map_tp, clock, SceneTransToPoint, DungeonEntryInfo, PersonalSceneJump, PlayerEnterDungeon
+from game_server.handlers import auth, ping, scene, avatar, entity, map_tp, clock
 
 if __name__ == "__main__":
     gameserver = GameServer("localhost", 22102)
@@ -11,10 +11,6 @@ if __name__ == "__main__":
     gameserver.add(entity.router)
     gameserver.add(map_tp.router)
     gameserver.add(clock.router)
-    gameserver.add(SceneTransToPoint.router)
-    gameserver.add(DungeonEntryInfo.router)
-    gameserver.add(PersonalSceneJump.router)
-    gameserver.add(PlayerEnterDungeon.router)
 
     gameserver.start()
     run_http_server('0.0.0.0')
